memdev/mockfs: widen base address fields to 64-bit
The mi_base field in dk_memdev_info_t and mockfs_memdev_base were
uint32_t, truncating physical addresses above 4GB. Widen to uint64_t
to match the actual page number range from mdevadd.

+++ bsd/dev/memdev.c
@@ -450,7 +450,7 @@ mdevioctl(dev_t dev, u_long cmd, caddr_t data, __unused int flag,
 		}
 		memdev_info->mi_mdev = TRUE;
 		memdev_info->mi_phys = (mdev[devid].mdFlags & mdPhys) ? TRUE : FALSE;
-		memdev_info->mi_base = (uint32_t)mdev[devid].mdBase;
+		memdev_info->mi_base = mdev[devid].mdBase;
 		memdev_info->mi_size = mdev[devid].mdSize;
 		break;
